+++ tests/test_CreateAthlete.py
@@ -32,7 +32,7 @@ def test_CreateAthletes_file(mock_post):
     mock_post.return_value = MagicMock(status_code=200, json=lambda: mock_success_response(formatted_time))
 
     # Authenticate
-    AuthManager(authMethod="file", env_file_name="tests/.env")
+    AuthManager(authMethod="file", env_file_name=r"tests/.env")
 
     # Create New Athletes
     players = [
@@ -102,7 +102,7 @@ def test_CreateAthletes_failure(mock_post):
     mock_post.return_value = MagicMock(status_code=200, json=lambda: mock_failure_response(formatted_time))
 
     # Authenticate
-    AuthManager(authMethod="file", env_file_name="tests/.env")
+    AuthManager(authMethod="file", env_file_name=r"tests/.env")
 
     # Create New Athletes
     players = [