+++ tests/test_GetMetrics.py
@@ -7,13 +7,12 @@ import pandas as pd
 def test_GetMetrics_file():
     
     # Authenticate
-    AuthManager(authMethod= "file", env_file_name= "tests/.env")
+    AuthManager(authMethod= "file", env_file_name= r"tests/.env")
     # Call for metrics
     metrics = GetMetrics()
     
     # Check response is DataFrame
     assert isinstance(metrics, pd.DataFrame)
-    assert isinstance(metrics.attrs['Count'], int)
 
 
 # successful call with env
@@ -26,4 +25,3 @@ def test_GetMetrics_env():
 
     # Check response is DataFrame
     assert isinstance(metrics, pd.DataFrame)
-    assert isinstance(metrics.attrs['Count'], int)