+++ tests/test_GetTypes.py
@@ -7,13 +7,12 @@ import pandas as pd
 def test_GetTypes_file():
 
     # Authenticate
-    AuthManager(authMethod= "file", env_file_name= "tests/.env")
+    AuthManager(authMethod= "file", env_file_name= r"tests/.env")
     # Call for test types
     data = GetTypes()
 
     # Check response is DataFrame
     assert isinstance(data, pd.DataFrame)
-    assert isinstance(data.attrs['Count'], int)
 
 
 # successful call with env
@@ -26,4 +25,3 @@ def test_GetTypes_env():
 
     # Check response is DataFrame
     assert isinstance(data, pd.DataFrame)
-    assert isinstance(data.attrs['Count'], int)